pyproject.toml
[tool.poetry]
name = "codenameseauserlogin"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
scratch2py = "^0.6.8"
mysql-connector-python = "^8.0.28"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
